all files in the norme
