+++ sources/tokenizer/add_token.c
@@ -6,7 +6,7 @@
 /*   By: phemsi-a <phemsi-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/22 21:11:55 by phemsi-a          #+#    #+#             */
-/*   Updated: 2021/07/10 10:53:38 by phemsi-a         ###   ########.fr       */
+/*   Updated: 2021/07/18 21:57:16 by phemsi-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,9 +60,7 @@ void	add_token(char *line, int start, int end, t_token **token_lst)
 	value = ft_substr(line, start, (end - start));
 	if (!value)
 		return ;
-	if (is_single_quote(value[0]))
-		remove_quotes(&value, SINGLE_QUOTE);
-	else if (ft_strchr(value, SINGLE_QUOTE) || ft_strchr(value, DOUBLE_QUOTE))
+	if (ft_strchr(value, SINGLE_QUOTE) || ft_strchr(value, DOUBLE_QUOTE))
 		check_quotes(&value);
 	else if (ft_strchr(value, '$'))
 		expand_variables(&value);
