+++ sources/tokenizer/define_type.c
@@ -6,7 +6,7 @@
 /*   By: phemsi-a <phemsi-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/10 10:53:17 by phemsi-a          #+#    #+#             */
-/*   Updated: 2021/07/10 10:56:30 by phemsi-a         ###   ########.fr       */
+/*   Updated: 2021/07/18 21:30:30 by phemsi-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,25 +15,25 @@
 static bool	is_builtin(char *value)
 {
 	if (!(ft_strcmp(value, "echo\0")) || !(ft_strcmp(value, "cd\0")))
-		return TRUE;
+		return (TRUE);
 	if (!(ft_strcmp(value, "pwd")) || !(ft_strcmp(value, "export")))
-		return TRUE;
+		return (TRUE);
 	if (!(ft_strcmp(value, "unset")) || !(ft_strcmp(value, "env")))
-		return TRUE;
+		return (TRUE);
 	if (!(ft_strcmp(value, "exit")))
-		return TRUE;
-	return FALSE;
+		return (TRUE);
+	return (FALSE);
 }
 
 static bool	is_operator(char *value)
 {
 	if (!(ft_strcmp(value, "|")) || !(ft_strcmp(value, "=")))
-		return TRUE;
+		return (TRUE);
 	if (!(ft_strcmp(value, ">")) || !(ft_strcmp(value, "<")))
-		return TRUE;
+		return (TRUE);
 	if (!(ft_strcmp(value, ">>")) || !(ft_strcmp(value, "<<")))
-		return TRUE;
-	return FALSE;
+		return (TRUE);
+	return (FALSE);
 }
 
 void	define_type(char *value, int *type)
