+++ sources/tokenizer/quotes.c
@@ -6,7 +6,7 @@
 /*   By: phemsi-a <phemsi-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/23 23:11:12 by phemsi-a          #+#    #+#             */
-/*   Updated: 2021/07/06 19:14:05 by phemsi-a         ###   ########.fr       */
+/*   Updated: 2021/07/18 21:31:55 by phemsi-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,31 +19,31 @@ static bool	has_a_match(char *line, int i, int *match_index, char c)
 		if (line[i] == c)
 		{
 			*match_index = i;
-			return TRUE;
+			return (TRUE);
 		}
 	}
-	return FALSE;
+	return (FALSE);
 }
 
 static bool	is_quote(char c)
 {
-	if (c == SINGLE_QUOTE|| c == DOUBLE_QUOTE)
-		return TRUE;
-	return FALSE;
+	if (c == SINGLE_QUOTE || c == DOUBLE_QUOTE)
+		return (TRUE);
+	return (FALSE);
 }
 
 bool	is_single_quote(char c)
 {
 	if (c == SINGLE_QUOTE)
-		return TRUE;
-	return FALSE;
+		return (TRUE);
+	return (FALSE);
 }
 
 bool	is_between_quotes(char *line, int i, int *match_index, char c)
 {
 	if (is_quote(c) && has_a_match(line, i, match_index, c))
-		return TRUE;
-	return FALSE;
+		return (TRUE);
+	return (FALSE);
 }
 
 void	remove_quotes(char **value, char quote)
